+++ src/core/schemas/stats.py
@@ -3,5 +3,5 @@ from pydantic import BaseModel
 
 class AddStatisticsData(BaseModel):
     symbols_per_minute: int
-    accuracy_percentage: int
+    accuracy_percentage: float
     mode_type: str
